refactor: stop bounding how many flags a deduper tracks
A record per flag, in each environment it is evaluated in, is the flag set
the environments serve, which LaunchDarkly already bounds. Evicting from
it only cost the hook a suppression it should have had, so the map is now
a plain HashMap and the record for a flag is updated in place.
